feat: pr comment refactos
